Core - Move LimitedConcurrencyLevelTaskScheduler to CefSharp namespace
In preparation for generating Ref assemblies which will expose some of our internal classes at least initially.
Move into our own namespace to avoid any conflict
