+++ CefSharp/Internals/MethodRunnerQueue.cs
@@ -5,7 +5,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Threading.Tasks.Schedulers;
+using CefSharp.Internals.Tasks;
 
 namespace CefSharp.Internals
 {
